[data] changed type from Int to Double of totalVolume field in CoinApiResponse

+++ data/src/main/java/com/davidepani/data/models/CoinApiResponse.kt
@@ -47,5 +47,5 @@ data class CoinApiResponse(
     @SerializedName("total_supply")
     val totalSupply: Double,
     @SerializedName("total_volume")
-    val totalVolume: Int
+    val totalVolume: Double
 )
